refactor: reuse dummy pid fixture
